refactor(platform): split HAL/Semaphore.h into Platform/<OS>/<OS>Semaphore.h (B1)
