+++ OloEngine/src/CMakeLists.txt
@@ -1040,6 +1040,7 @@
 if(WIN32)
 	list(APPEND SOURCES
 		"Platform/Windows/WindowsEvent.h"
+		"Platform/Windows/WindowsSemaphore.h"
 		"Platform/Windows/WindowsInput.cpp"
 		"Platform/Windows/WindowsWindow.cpp"
 		"Platform/Windows/WindowsWindow.h"
@@ -1048,6 +1049,7 @@ if(WIN32)
 elseif(UNIX AND NOT APPLE)
 	list(APPEND SOURCES
 		"Platform/Linux/LinuxEvent.h"
+		"Platform/Linux/LinuxSemaphore.h"
 		"Platform/Linux/LinuxInput.cpp"
 		"Platform/Linux/LinuxWindow.cpp"
 		"Platform/Linux/LinuxWindow.h"
